SelfLevelControl:  - Added typedef for the throttle and angle types.  - Changed throttle to instead use altitude extension instead of y-acceleration. PIDControl:  - Added cast for time to double. ControlScheme:  - Removed all dependence of Controller and SensorRepository. This class is now (and should stay in the future) as a interface class for the different control schemes. Controller:  - Renamed PIDType->AngleType, and added a ThrottleType template argument.  - Changed to using units::unit_cast for casting due to now using ThrottleType instead of meters_per_second_squared_t.

+++ Src/Control/PIDControl.hpp
@@ -24,7 +24,7 @@ namespace Copter::Control
             const Type diffError = error - mPrevError;
 
             // Calculate time difference
-            double curTime = Kernel::Clock::now().time_since_epoch().count();
+            double curTime = static_cast<double>(Kernel::Clock::now().time_since_epoch().count());
             double diffTime = (curTime - mPrevTime) / 1000;
             mPrevTime = curTime;
